Fix warnings on MSVC
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@32513 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ plugins/video/H.263-1998/rfc2190.cxx
@@ -63,7 +63,7 @@ static size_t FindByteAlignedCode(const unsigned char * base, size_t len, const 
 
 static int FindPSC(const unsigned char * base, size_t len)
 {
-  return FindByteAlignedCode(base, len, PSC, PSC_Mask, sizeof(PSC));
+  return (int)FindByteAlignedCode(base, len, PSC, PSC_Mask, sizeof(PSC));
 }
 
 #if 0
